fix: Fix some bugs

+++ src/main/java/de/aivot/GoverBackend/controllers/DestinationController.java
@@ -2,6 +2,7 @@ package de.aivot.GoverBackend.controllers;
 
 import de.aivot.GoverBackend.models.entities.Destination;
 import de.aivot.GoverBackend.permissions.IsAdmin;
+import de.aivot.GoverBackend.repositories.ApplicationRepository;
 import de.aivot.GoverBackend.repositories.DestinationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,10 +16,12 @@ import java.util.Optional;
 @RestController
 public class DestinationController {
     private final DestinationRepository repository;
+    private final ApplicationRepository applicationRepository;
 
     @Autowired
-    public DestinationController(DestinationRepository repository) {
+    public DestinationController(DestinationRepository repository, ApplicationRepository applicationRepository) {
         this.repository = repository;
+        this.applicationRepository = applicationRepository;
     }
 
     @GetMapping("/api/destinations")
@@ -79,6 +82,10 @@ public class DestinationController {
         Optional<Destination> optDest = repository.findById(id);
         if (optDest.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+
+        if (applicationRepository.existsByDestination_Id(id)) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT);
         }
 
         repository.delete(optDest.get());
